fix avro artifact type detection
