+++ app/src/main/java/io/apicurio/registry/rest/v2/GroupsResourceImpl.java
@@ -520,7 +520,7 @@ public class GroupsResourceImpl implements GroupsResource {
                 content = ContentTypeUtil.yamlToJson(content);
             }
 
-            ArtifactType artifactType = determineArtifactType(content, xRegistryArtifactType, ct);
+            ArtifactType artifactType = ArtifactTypeUtil.determineArtifactType(content, xRegistryArtifactType, ct);
             rulesService.applyRules(gidOrNull(groupId), artifactId, artifactType, content, RuleApplicationType.CREATE);
             final String finalArtifactId = artifactId;
             EditableArtifactMetaDataDto metaData = getEditableMetaData(artifactName, artifactDescription);
@@ -638,58 +638,6 @@ public class GroupsResourceImpl implements GroupsResource {
             return null;
         }
         return new String(Base64.decode(encoded));
-    }
-
-    /**
-     * Figures out the artifact type in the following order of precedent:
-     * <p>
-     * 1) The provided X-Registry-ArtifactType header
-     * 2) A hint provided in the Content-Type header
-     * 3) Determined from the content itself
-     *
-     * @param content       the content
-     * @param xArtifactType the artifact type
-     * @param ct            content type from request API
-     */
-    private static ArtifactType determineArtifactType(ContentHandle content, ArtifactType xArtifactType, String ct) {
-        ArtifactType artifactType = xArtifactType;
-        if (artifactType == null) {
-            artifactType = getArtifactTypeFromContentType(ct);
-            if (artifactType == null) {
-                artifactType = ArtifactTypeUtil.discoverType(content, ct);
-            }
-        }
-        return artifactType;
-    }
-
-    /**
-     * Tries to figure out the artifact type by analyzing the content-type.
-     *
-     * @param contentType the content type header
-     */
-    private static ArtifactType getArtifactTypeFromContentType(String contentType) {
-        if (contentType != null && contentType.contains(MediaType.APPLICATION_JSON) && contentType.indexOf(';') != -1) {
-            String[] split = contentType.split(";");
-            if (split.length > 1) {
-                for (String s : split) {
-                    if (s.contains("artifactType=")) {
-                        String at = s.split("=")[1];
-                        try {
-                            return ArtifactType.valueOf(at);
-                        } catch (IllegalArgumentException e) {
-                            throw new BadRequestException("Unsupported artifact type: " + at);
-                        }
-                    }
-                }
-            }
-        }
-        if (contentType != null && contentType.contains("x-proto")) {
-            return ArtifactType.PROTOBUF;
-        }
-        if (contentType != null && contentType.contains("graphql")) {
-            return ArtifactType.GRAPHQL;
-        }
-        return null;
     }
 
     private ArtifactMetaData handleIfExists(String groupId, String artifactId, String version,
